toDocker

+++ src/upload/main.c
@@ -109,9 +109,9 @@ printCacheStats()
 {
   printf("Cache Statistics:\n");
   if (icacheSets) {
-    printf("  total I-cache accesses:  %10lu\n", (unsigned long)icacheRefs);
-    printf("  total I-cache misses:    %10lu\n", (unsigned long)icacheMisses);
-    printf("  total I-cache penalties: %10lu\n", (unsigned long)icachePenalties);
+    printf("  total I-cache accesses:  %10lu\n", icacheRefs);
+    printf("  total I-cache misses:    %10lu\n", icacheMisses);
+    printf("  total I-cache penalties: %10lu\n", icachePenalties);
     if (icacheRefs > 0) {
       printf("  I-cache miss rate:   %17.2f%%\n",
           100.0*(double)icacheMisses/(double)icacheRefs);
@@ -123,9 +123,9 @@ printCacheStats()
     }
   }
   if (dcacheSets) {
-    printf("  total D-cache accesses:  %10lu\n", (unsigned long)dcacheRefs);
-    printf("  total D-cache misses:    %10lu\n", (unsigned long)dcacheMisses);
-    printf("  total D-cache penalties: %10lu\n", (unsigned long)dcachePenalties);
+    printf("  total D-cache accesses:  %10lu\n", dcacheRefs);
+    printf("  total D-cache misses:    %10lu\n", dcacheMisses);
+    printf("  total D-cache penalties: %10lu\n", dcachePenalties);
     if (dcacheRefs > 0) {
       printf("  D-cache miss rate:   %17.2f%%\n",
           100.0*(double)dcacheMisses/(double)dcacheRefs);
@@ -137,9 +137,9 @@ printCacheStats()
     }
   }
   if (l2cacheSets) {
-    printf("  total L2-cache accesses: %10lu\n", (unsigned long)l2cacheRefs);
-    printf("  total L2-cache misses:   %10lu\n", (unsigned long)l2cacheMisses);
-    printf("  total L2-cache penalties:%10lu\n", (unsigned long)l2cachePenalties);
+    printf("  total L2-cache accesses: %10lu\n", l2cacheRefs);
+    printf("  total L2-cache misses:   %10lu\n", l2cacheMisses);
+    printf("  total L2-cache penalties:%10lu\n", l2cachePenalties);
     if (l2cacheRefs > 0) {
       printf("  L2-cache miss rate:  %17.2f%%\n",
           100.0*(double)l2cacheMisses/(double)l2cacheRefs);
@@ -243,8 +243,8 @@ main(int argc, char *argv[])
   printStudentInfo();
   printCacheConfig();
   printCacheStats();
-  printf("Total Memory accesses:  %lu\n", (unsigned long)totalRefs);
-  printf("Total Memory penalties: %lu\n", (unsigned long)totalPenalties);
+  printf("Total Memory accesses:  %lu\n", totalRefs);
+  printf("Total Memory penalties: %lu\n", totalPenalties);
   if (totalRefs > 0) {
     printf("avg Memory access time: %13.2f cycles\n",
         (double)totalPenalties / totalRefs);
